==> scripts/0-full-analysis.R <==
# run analysis in order
source("scripts/1-preprocessing.R")
source("scripts/2-integrated-clustering.R")
source("scripts/3-annotation.R")
source("scripts/4-basic-figures.R")
source("scripts/5-niche-markers.R")
source("scripts/6-genes-of-interest.R")
source("scripts/7-signatures-of-interest.R")
source("scripts/8-differential-abundance.R")
source("scripts/9-differential-gene-expression.R")
source("scripts/10-ora.R")
source("scripts/11-spatially-variable-features.R")
